+++ server/routes/join.js
@@ -4,36 +4,42 @@ const db = require('../lib/db');
 const qs = require('querystring');
 
 
-router.post('/', async function (req, res) { //회원가입
-  const post = req.body;
-  let checkId = false;
-  let checkNickname = false;
-  await db.query(`SELECT user_id FROM user where user_id=?`,[post.id], function (error, result) {
-    if (error) {
-      throw error;
-    }
-    if(result.length==0){
+module.exports = function(passport){
+  router.get('/id', async function(req,res){ //id 중복확인
+    const post = req.body;
+    await db.query(`SELECT user_id FROM user where user_id=?`,[post.id], function (error, result) {
+      if (error) {
+        throw error;
+      }
+      if(result.length==0){
         console.log('id ok');
-        checkId=true;
-        console.log(checkId);
-    }
+        res.json({already: false});  //중복없으면
+      } else{
+        res.json({already: true});  //중복있으면
+      }
+    });
   });
 
-  await db.query(`SELECT nickname FROM user where nickname=?`,[post.nickname], function (error, result) {
-    if (error) {
-      throw error;
-    }
-    if(result.length==0){
-        console.log('nickname ok');
-        checkNickname=true;
-        console.log(checkNickname);
-        console.log(post);
-    }
+  router.get('/nick', async function(req,res){ //nickname 중복확인
+    const post = req.body;
+    await db.query(`SELECT user_id FROM user where user_id=?`,[post.id], function (error, result) {
+      if (error) {
+        throw error;
+      }
+      if(result.length==0){
+        console.log('id ok');
+        res.json({already: false});  //중복없으면
+      } else{
+        res.json({already: true});  //중복있으면
+      }
+    });
   });
 
-  if (checkId && checkNickname) {
-    console.log("insert start");
-    
+
+  router.post('/', async function (req, res) { //회원가입
+    const post = req.body;
+    let user;
+      
     await db.query(
       `INSERT INTO user(user_id,password,nickname,gender) 
             VALUES(?,?,?,?);`,
@@ -46,7 +52,9 @@ router.post('/', async function (req, res) { //회원가입
         console.log("insert complete");
       }
     );
+
     //회원가입 성공 시 바로 로그인.
+    
     await db.query(
       `SELECT * FROM user WHERE user_id=?`,[post.id],
       function(error,result){
@@ -54,19 +62,23 @@ router.post('/', async function (req, res) { //회원가입
           console.log('mysql err');
           throw error;
         }
-        let user = result[0];
+        user = result[0];
       }
     );
     req.login(user, function(err){ //회원가입 성공 시 바로 로그인.
       return res.send('ok');
     });
-  } else {
-    if (!checkId) {
-      res.status(400).send({ error: 'Already used id' });
-    } else {
-      res.status(400).send({ error: 'Already used Nickname' });
+    
+  /*
+      else {
+      if (!checkId) {
+        res.status(400).send({ error: 'Already used id' });
+      } else {
+        res.status(400).send({ error: 'Already used Nickname' });
+      }
     }
-  }
-});
+  */
 
-module.exports = router;
+  });
+  return router;
+}
